+++ R/plot_gpt.R
@@ -229,8 +229,8 @@ gpt_helper <- function(
 
   # final data frame for plot
   newdf <- data.frame()
-  group.p <- c()
-  group.n <- c()
+  group.p <- character()
+  group.n <- character()
 
   # create data frame, for dplyr-chain
   mydf <-
@@ -261,11 +261,11 @@ gpt_helper <- function(
     dplyr::summarise(N = n(), p = suppressWarnings(stats::chisq.test(table(.data$xpos, .data$dep))$p.value))
 
   # copy p values
-  for (i in seq_len(length(pvals$grp))) group.p <- c(group.p, get_p_stars(pvals$p[i]))
+  for (i in seq_len(length(pvals$grp))) group.p[i] <- get_p_stars(pvals$p[i])
 
   # copy N
   for (i in seq_len(length(pvals$grp)))
-    group.n <- c(group.n, prettyNum(pvals$N[i], big.mark = ",", scientific = F))
+    group.n[i] <- prettyNum(pvals$N[i], big.mark = ",", scientific = F)
 
   # if we want total line, repeat all for
   # complete data frame