+++ NEWS.md
@@ -19,6 +19,7 @@ Following functions are now defunct:
 * `plot_model()` does no longer automatically change the plot-type to `"slope"` for models with only one predictor that is categorical and has more than two levels.
 * `type = "eff"` and `type = "pred"` in `plot_model()` did not when `terms` was not specified. 
 * If robust standard errors are requested in `tab_model()`, the confidence intervals and p-values are now re-calculated and adjusted based on the robust standard errors.
+* `colors = "bw"` was not recognized correctly for `plot_model(..., type = "int")`.
 
 # sjPlot 2.6.1