+++ NEWS.md
@@ -3,6 +3,7 @@
 ## General
 
 * Revised some help-files and vignettes.
+* Replaced package dependency **broom** with **generics**.
 
 ## Removed / Defunct