+++ NEWS.md
@@ -1,3 +1,9 @@
+# sjPlot 2.6.3
+
+## Bug fixes
+
+* Legend labels were inverted for **brms**-models in `plot_model(..., type = "diag")`.
+
 # sjPlot 2.6.2
 
 ## General